# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DDRI API
  description: 비로그인 조회형 DDRI 웹서비스 API
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /v1/weather/direct:
    get:
      tags: [weather]
      summary: 주간 날씨 조회
      description: Open-Meteo에서 일별 예보 배열을 조회한다.
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            example: 37.5665
        - name: lon
          in: query
          required: true
          schema:
            type: number
            example: 126.978
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-20"
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherForecastResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/weather/direct/single:
    get:
      tags: [weather]
      summary: 특정 시각 날씨 조회
      description: 선택 날짜 또는 기준 시각에 가장 가까운 상세 날씨 1건을 조회한다.
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            example: 37.5665
        - name: lon
          in: query
          required: true
          schema:
            type: number
            example: 126.978
        - name: target_datetime
          in: query
          required: false
          schema:
            type: string
            example: "2026-03-20T18:00:00+09:00"
        - name: target_date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-20"
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherSingleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  schemas:
...
    WeatherForecastResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/WeatherDailyItem"
    WeatherSingleResponse:
      type: object
      properties:
        result:
          oneOf:
            - $ref: "#/components/schemas/WeatherSingleItem"
            - type: string
              example: Error
        errorMsg:
          type: string
          nullable: true
...
    WeatherDailyItem:
      type: object
      properties:
        weather_datetime:
          type: string
        weather_type:
          type: string
        weather_low:
          type: number
        weather_high:
          type: number
        precipitation_probability_max:
          type: number
        icon_url:
          type: string
    WeatherSingleItem:
      type: object
      properties:
        weather_datetime:
          type: string
        weather_type:
          type: string
        weather_low:
          type: number
        weather_high:
          type: number
        temperature:
          type: number
        precipitation_probability:
          type: number
        icon_url:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"